src/JuegoFachadaTest.java
import org.junit.Before;
import org.junit.Test;

import java.util.Date;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;

public class JuegoFachadaTest {
    private JuegoFachada juegoFachada;
    Juego juego;
    Date fecha;
    @Before
    public void setUp() {
        // Aquí puedes inicializar tu juegoFachada y configurar las listas de juegos y usuarios para las pruebas.
        juegoFachada = JuegoFachadaImpl.getInstance();
        // Configura las listas de juegos y usuarios según tus necesidades.
    }

    @Test
    public void testIniciarPartida() {
        // Prueba iniciarPartida con un juego y usuario existentes.
        // Asegúrate de que los juegos y usuarios estén configurados en el entorno de prueba.
        juegoFachada.iniciarPartida("1a", "2b");
        // Añade aserciones para verificar el resultado esperado.
    }

    @Test
    public void testConsultarNivel() {
        // Prueba consultarNivel con un usuario existente.
        // Asegúrate de que los usuarios estén configurados en el entorno de prueba.
        int nivel = juegoFachada.consultarNivel("1a");
        // Añade aserciones para verificar el resultado esperado.
    }

    @Test
    public void testConsultarPuntuacion() {
        // Prueba consultarPuntuacion con un usuario existente.
        // Asegúrate de que los usuarios estén configurados en el entorno de prueba.
        int puntuacion = juegoFachada.consultarPuntuacion("1a");
        // Añade aserciones para verificar el resultado esperado.
    }

    @Test
    public void testPasarDeNivel() {
        // Prueba PasarDeNivel con un usuario existente.
        // Asegúrate de que los usuarios estén configurados en el entorno de prueba.
        juegoFachada.PasarDeNivel("1a", 15, fecha);
        // Añade aserciones para verificar el resultado esperado.
    }

    @Test
    public void testFinalizarJuego() {
        // Prueba finalizarJuego con un usuario existente.
        // Asegúrate de que los usuarios estén configurados en el entorno de prueba.
        juegoFachada.finalizarJuego("1a");
        // Añade aserciones para verificar el resultado esperado.
    }
}